Turns out we may not need a username or password either
The ssl reject unauthorized option might make it so that we don't have to worry about the rotating usernames and passwords from earlier. At least, that's if my testing is at all conclusive, which it may not be. I only tested SELECT queries, after all. Still, the original code is still there, just commented.